refactor(effects): split advanced screen-space passes
