+++ modules/effects/src/passes/postprocessing/advanced/screen-space-shader-helpers.ts
@@ -1,0 +1,23 @@
+// luma.gl
+// SPDX-License-Identifier: MIT
+// Copyright (c) vis.gl contributors
+
+/** Shared WGSL helpers for depth-based screen-space effects. */
+export const depthHelpers = /* wgsl */ `\
+fn advancedSceneUV(uv: vec2f) -> vec2f {
+  return uv;
+}
+
+fn advancedLinearDepth(depth: f32, nearPlane: f32, farPlane: f32) -> f32 {
+  return (nearPlane * farPlane) / max(farPlane - depth * (farPlane - nearPlane), 0.0001);
+}
+
+fn advancedDepthNormal(depthTexture: texture_depth_2d, depthTextureSampler: sampler, uv: vec2f) -> vec3f {
+  let dimensions = vec2i(textureDimensions(depthTexture));
+  let coordinate = clamp(vec2i(uv * vec2f(dimensions)), vec2i(0), dimensions - vec2i(1));
+  let center = textureLoad(depthTexture, coordinate, 0);
+  let right = textureLoad(depthTexture, min(coordinate + vec2i(1, 0), dimensions - vec2i(1)), 0);
+  let up = textureLoad(depthTexture, min(coordinate + vec2i(0, 1), dimensions - vec2i(1)), 0);
+  return normalize(vec3f((center - right) * f32(dimensions.x), (center - up) * f32(dimensions.y), 1.0));
+}
+`;
